# server/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Teach RTC",
    "description": "A sample application showing multi-user editing.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "parameters": {
      "DocumentUuid": {
        "description": "Document universally unique identifier",
        "in": "path",
        "name": "documentUuid",
        "schema": {
          "$ref": "#/components/schemas/DocumentUuid"
        },
        "required": true
      },
      "Username": {
        "description": "User name",
        "in": "path",
        "name": "username",
        "schema": {
          "$ref":"#/components/schemas/Username"
        },
        "required": true
      },
...
    },
    "responses": {
      "LoginSuccess": {
        "description": "Successfully created JWT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginResp"
            }
          }
        }
      },
      "LoginValidation": {
        "description": "Inputs failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/InvalidUsername"
                },
                {
                  "$ref": "#/components/schemas/InvalidPassword"
                }
              ]
            }
          }
        }
      },
      "LoginFailure": {
        "description": "Username or password incorrect.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginFailure"
            }
          }
        }
      },
      "ListUsersSuccess": {
        "description": "Successfully got the list of users.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ListUsersResp"
            }
          }
        }
      },
      "CreateUserSuccess": {
        "description": "Successfully created a user.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateUserResp"
            }
          }
        }
      },
      "GetUserSuccess": {
        "description": "Successfully got details about a user.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GetUserResp"
            }
          }
        }
      },
      "CreateUserValidation": {
        "description": "Inputs failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/InvalidUsername"
                },
                {
                  "$ref": "#/components/schemas/InvalidPassword"
                },
                {
                  "$ref": "#/components/schemas/InvalidFullName"
                }
              ]
            }
          }
        }
      },
      "CreateUserConflict": {
        "description": "Username is already in use.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateUserConflict"
            }
          }
        }
      },
      "GetUserNotFound": {
        "description": "The user was not found in the database.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GetUserNotFound"
            }
          }
        }
      },
      "ListDocumentsSuccess": {
        "description": "Successfully listed documents.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ListDocumentsResp"
            }
          }
        }
      },
      "CreateDocumentSuccess": {
        "description": "Successfully created a new document.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateDocumentResp"
            }
          }
        }
      },
      "CreateDocumentValidation": {
        "description": "Inputs failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/InvalidTitle"
            }
          }
        }
      },
      "GetDocumentTitleSuccess": {
        "description": "Success",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GetDocumentTitleResp"
            }
          }
        }
      },
...
      "AuthorizationCheckFailed": {
        "description": "The authorization check failed due to a missing or invalid JWT bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthorizationCheckFailed"
            }
          }
        }
      },
      "DocumentUnknownOrForbidden": {
        "description": "User is not permitted to access this document in this way (or it doesn't exist).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DocumentUnknownOrForbidden"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "LoginReq": {
        "type": "object",
        "properties": {
          "username": {
            "$ref": "#/components/schemas/Username"
          },
          "password": {
            "$ref": "#/components/schemas/Password"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "additionalProperties": false
      },
...
      "CreateUserReq": {
        "type": "object",
        "properties": {
          "username": {
            "$ref": "#/components/schemas/Username"
          },
          "password": {
            "$ref": "#/components/schemas/Password"
          },
          "fullName": {
            "$ref": "#/components/schemas/FullName"
          }
        },
        "required": [
          "username",
          "password",
          "fullName"
        ],
        "additionalProperties": false
      },
...
      "CreateDocumentReq": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "My document"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
    "/jwt": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authenticate and create a JWT bearer token.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/LoginSuccess"
          },
          "400": {
            "$ref": "#/components/responses/LoginValidation"
          },
          "401": {
            "$ref": "#/components/responses/LoginFailure"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List all the users.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ListUsersSuccess"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserReq"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/CreateUserSuccess"
          },
          "400": {
            "$ref": "#/components/responses/CreateUserValidation"
          },
          "409": {
            "$ref": "#/components/responses/CreateUserConflict"
          }
        }
      }
    },
    "/users/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get details of a user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref":"#/components/parameters/Username"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetUserSuccess"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "404": {
            "$ref": "#/components/responses/GetUserNotFound"
          }
        }
      }
    },
    "/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List all the documents that can be seen by the current user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ListDocumentsSuccess"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Create a new document.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/CreateDocumentSuccess"
          },
          "400": {
            "$ref": "#/components/responses/CreateDocumentValidation"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          }
        }
      }
    },
    "/documents/{documentUuid}/title": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get document title.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentUuid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetDocumentTitleSuccess"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "403": {
            "$ref": "#/components/responses/DocumentUnknownOrForbidden"
          }
        }
      }
    },
    "/documents/{documentUuid}/jwt": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get a JSON Web Token to access the document.",
        "description": "Get a JSON Web Token that relates specifically to this document and the role the user has.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentUuid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetDocumentJWTSuccess"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "403": {
            "$ref": "#/components/responses/DocumentUnknownOrForbidden"
          }
        }
      }
    },
    "/documents/{documentUuid}/key": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get the secret key and key hint associated with the document.",
        "description": "Get the secret key and key hint associated with the document. The secret key is used by RTC to encrypt all content and messages that are sent to the RTC server. This ensures privacy as the RTC server has no way to read or recreate the content of the document. When no key hint is provided the integration is allowed to generate a new key with an associated unique key hint (for example the time of creation), however if a key hint is provided then the key that was originally issued with that hint must be returned.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentUuid"
          },
          {
            "$ref": "#/components/parameters/KeyHint"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetDocumentKeySuccess"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "403": {
            "$ref": "#/components/responses/DocumentUnknownOrForbidden"
          },
          "404": {
            "$ref": "#/components/responses/KeyUnknown"
          }
        }
      }
    },
    "/documents/{documentUuid}/content": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get document content.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentUuid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetDocumentContentSuccess"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "403": {
            "$ref": "#/components/responses/DocumentUnknownOrForbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Documents"
        ],
        "summary": "Save document content.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentUuid"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentContentReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UpdateDocumentContentSuccess"
          },
          "400": {
            "$ref": "#/components/responses/UpdateDocumentContentValidation"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "403": {
            "$ref": "#/components/responses/DocumentUnknownOrForbidden"
          }
        }
      }
    },
    "/documents/{documentUuid}/collaborators": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get all collaborators and their permissions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentUuid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetCollaboratorsSuccess"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "403": {
            "$ref": "#/components/responses/DocumentUnknownOrForbidden"
          }
        }
      }
    },
    "/documents/{documentUuid}/collaborators/{username}": {
      "put": {
        "tags": [
          "Documents"
        ],
        "summary": "Change collaborator permissions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentUuid"
          },
          {
            "$ref":"#/components/parameters/Username"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAccessReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SetAccessSuccess"
          },
          "400": {
            "$ref": "#/components/responses/SetAccessValidation"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "403": {
            "$ref": "#/components/responses/DocumentUnknownOrForbidden"
          }
        }
      }
    },
    "/users/{username}/documents/{documentUuid}/access": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get the collaborators access level",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref":"#/components/parameters/Username"
          },
          {
            "$ref": "#/components/parameters/DocumentUuid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetCollaboratorsAccessSuccess"
          },
          "400": {
            "$ref": "#/components/responses/SetAccessValidation"
          },
          "401": {
            "$ref": "#/components/responses/AuthorizationCheckFailed"
          },
          "403": {
            "$ref": "#/components/responses/DocumentUnknownOrForbidden"
          }
        }
      }
    }
  }
}